+++ app/src/main/java/com/karl/android/coincounter/settings/SettingsActivity.java
@@ -1,4 +1,4 @@
-package com.karl.android.coincounter;
+package com.karl.android.coincounter.settings;
 
 
 import android.annotation.TargetApi;
@@ -20,6 +20,8 @@ import android.support.v4.app.NavUtils;
 import android.support.v7.app.ActionBar;
 import android.text.TextUtils;
 import android.view.MenuItem;
+
+import com.karl.android.coincounter.R;
 
 import java.util.List;