+++ app/src/main/java/com/karl/android/fragments/MySavesFragment.java
@@ -6,8 +6,10 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ListView;
+import android.widget.TextView;
 
 import com.karl.android.coincounter.R;
+import com.karl.android.templates.SlimSave;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -19,11 +21,9 @@ public class MySavesFragment extends Fragment {
     View rootView;
 
     ListView savesList;
+    TextView noInformationToDisplay;
 
-    public MySavesFragment() {
-        // Required empty public constructor
-    }
-
+    public MySavesFragment() {}
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
@@ -33,6 +33,8 @@ public class MySavesFragment extends Fragment {
         // initialise the views of the fragment.
         initialiseViews();
 
+        savesList.setVisibility(View.GONE);
+
         return rootView;
     }
 
@@ -40,6 +42,25 @@ public class MySavesFragment extends Fragment {
      * Initialise the views for the fragment.
      */
     private void initialiseViews(){
+        // lists
         savesList = (ListView) rootView.findViewById(R.id.my_saves_fragment_list_view);
+
+        // text views
+        noInformationToDisplay = (TextView) rootView.findViewById(R.id.my_saves_fragment_no_information_text_view);
+    }
+
+    /**
+     * This method gets the information from the database and return it in a Save object to be added into the views.
+     * This method only returns the essential information that the app needs in order to display
+     * the list of items for the user to view.
+     *
+     * @return a SlimSave object with the information from the database.
+     */
+    private SlimSave getInformationFromDatabase(){
+        SlimSave save = new SlimSave();
+
+        // get information from the database and add it to the save object.
+
+        return save;
     }
 }