+++ changelog.md
@@ -2,7 +2,47 @@
 # Changelog
 
 # 15
-- Removed all languages but English
+- Removed support for Afrikaans
+- Removed support for Bulgarian
+- Removed support for Czech
+- Removed support for Danish
+- Removed support for German
+- Removed support for Greek
+- Removed support for Esperanto
+- Removed support for Spanish
+- Removed support for Estonian
+- Removed support for Finnish
+- Removed support for Irish
+- Removed support for Hebrew
+- Removed support for Croatian
+- Removed support for Hungarian
+- Removed support for Lithuanian
+- Removed support for Latvian
+- Removed support for Maltese
+- Removed support for Dutch
+- Removed support for Polish
+- Removed support for Portuguese
+- Removed support for Romanian
+- Removed support for Russian
+- Removed support for Solvak
+- Removed support for Slovenian
+- Removed support for Swedish
+- Removed support for Ukrainian
+- Removed support for Italian
+- Removed support for Icelandic
+- Removed support for Belarusian
+- Removed support for Bosnian
+- Removed support for Japanese
+- Removed support for Korean
+- Removed support for Norwegian
+- Removed support for Turkish
+- Removed support for Albanian
+- Removed support for Serbian
+- Removed support for Chinese
+- Removed support for Thai
+- Removed support for Catalan
+- Removed support for Basque
+- Removed support for Filipino
 * Moved MySQLHelper to new package com.karl.android.data.sql
 + Added com.karl.android.fragments
 + Added MySavesFragment
@@ -15,3 +55,4 @@
 + Denominations are now saved into the database
 + Added settings activity 
 * Reduced the splash screen from 2.5 to .5 seconds 
++ Added SlimSave as a minimum requirement Save object